Refactor WebP subscriber
